+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v1.0.2
+Add functionalities for working with ticket tasks:
+- Get a task by app ID, ticket ID and task ID.
+- Edit a task by app ID, ticket ID, task ID, and edited task data
+
 ## v1.0.1
 Add functionalities for working with KnowledgeBase articles:
 - Get an article by ID.